research/sweng25_group22_multiagentsimframework/backend/src/requirements.txt
python-dotenv
ipdb
numpy
openai
autogen
autogen_ext
autogen_core
autogen_agentchat
pymongo
flask
flask-cors
prefect